add support for multichannel clip composed of multiple mono files
To do so, we can't access aafiAudioEssence from aafiAudioClip->essence 
anymore.

Insted, we use aafiAudioClip->essencePointerList, which hold a list of 
aafiAudioEssencePointer structures, each one holding a pointer to an 
essence (aafiAudioEssence).

This allows to have multiple aafiAudioEssence structs per clip (each one 
describing a file and composing one channel of the clip).

This also allows to reuse the same essence file in any number of clips.

- add struct aafiAudioEssencePointer
- add macro AAFI_foreachAudioEssencePointerInFile()
- add macro AAFI_foreachAudioEssencePointer()
- add function aafi_newAudioEssencePointer()
- add function aafi_freeAudioEssencePointer()
- add function aafi_getAudioEssencePointerChannelCount()